Ligação e config a base de dados - Migração efetuado - adicionado enums como apoio

+++ DAL_ExpenseManagement/DAL_ExpenseManagement/Models/Saving.cs
@@ -10,7 +10,7 @@ namespace DAL_ExpenseManagement.Models
     {
         public int Id { get; set; }
         public int UserId { get; set; }
-        public decimal Goal { get; set; }
+        public float Goal { get; set; }
 
         public User User { get; set; }
     }
